Tweaks to existing structure and code.
