+++ src/mfi_alm/assets/__init__.py
@@ -1,7 +1,9 @@
 from .asset import Asset
+from .asset_portfolio import AssetPortfolio
 from .bonds import FixedBond
 
 __all__ = [
     "Asset",
+    "AssetPortfolio",
     "FixedBond",
 ]
