+++ src/mfi_alm/assets/asset.py
@@ -1,7 +1,6 @@
-from .bonds import FixedBond
 from typing import Self
-import numpy as np
-from copy import deepcopy
+
+from mfi_alm.assets.bonds import FixedBond
 
 
 class Asset:
@@ -16,16 +15,3 @@ class Asset:
 
     def copy(self) -> Self:
         return Asset(fixed_bond=self.fixed_bond.copy(), ytm=self.ytm)
-
-
-class AssetPortfolio:
-
-    def __init__(self, assets: list[Asset], ytm: float):
-        self.assets = [a.copy() for a in assets]
-        self.ytm = ytm
-
-    def market_value(self) -> float:
-        return np.sum([a.market_value() for a in self.assets])
-
-    def copy(self) -> Self:
-        return AssetPortfolio(assets=deepcopy(self.assets), ytm=self.ytm)
