+++ src/mfi_alm/assets/asset_portfolio.py
@@ -1,0 +1,17 @@
+from typing import Self
+
+import numpy as np
+
+from mfi_alm.assets.asset import Asset
+
+
+class AssetPortfolio:
+    def __init__(self, assets: list[Asset], ytm: float):
+        self.assets = [a.copy() for a in assets]
+        self.ytm = ytm
+
+    def market_value(self) -> float:
+        return np.sum([a.market_value() for a in self.assets])
+
+    def copy(self) -> Self:
+        return AssetPortfolio(assets=self.assets.copy(), ytm=self.ytm)
