+++ src/mfi_alm/liabilities/__init__.py
@@ -1,11 +1,9 @@
-from .liability import Liability
 from .mortality import MortalityModel
 from .insurance import WholeLifeInsurance
 from .policyholder import Policyholder
 
 
 __all__ = [
-    "Liability",
     "MortalityModel",
     "WholeLifeInsurance",
     "Policyholder",
