--- src/mfi_alm/liabilities/liability.py
@@ -1,19 +1,0 @@
-"""
-TODO:
-Create a general liability class that contains the following attributes:
-price, cusip. Can leverage info from placeholder function.ipynb.
-@dataclass
-class Liability:
-    cost: float
-
-    def evolve_end_of_month(self, scenario: dict):
-        # Placeholder logic
-        pass
-
-Include the following methods:
-evolve_end_of_month
-"""
-
-
-class Liability:
-    pass
